apps/ppc-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-ads==23.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
cryptography==41.0.7
bingads==13.0.18
apscheduler==3.10.4
